move EventHandler code to a more logical location
