+++ src/main/java/telraam/beacon/TCPFactory.java
@@ -3,28 +3,20 @@ package telraam.beacon;
 import java.io.IOException;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.List;
-import java.util.ArrayList;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
-* The meat and potato's, but actually just spawning new connections with a creator,
-* and exposing subscriber pattern.
+* The meat and potato's, but actually just spawning new connections with a creator.
 *
 * @author  Arthur Vercruysse
 */
-public class TCPFactory<B> implements Event.EventHandler<B>, Runnable {
+public class TCPFactory<B> implements Runnable {
     private static Logger logger = Logger.getLogger(TCPFactory.class.getName());
 
     private ServerSocket socket;
     protected Callback<Void, Socket> creator;
-
-    protected List<Callback<Void, B>> handlers = new ArrayList<>();
-    protected List<Callback<Void, Exception>> errorHandlers = new ArrayList<>();
-    protected List<Callback<Void, Void>> exitHandlers = new ArrayList<>();
-    protected List<Callback<Void, Void>> connectHandlers = new ArrayList<>();
 
     public TCPFactory(Callback<Void, Socket> creator, int port) throws IOException {
         this(port);
@@ -47,41 +39,5 @@ public class TCPFactory<B> implements Event.EventHandler<B>, Runnable {
                 throw new RuntimeException(e);
             }
         }
-    }
-
-    public TCPFactory<B> onError(Callback<Void, Exception> handler) {
-        this.errorHandlers.add(handler);
-        return this;
-    }
-
-    public TCPFactory<B> onData(Callback<Void, B> handler) {
-        this.handlers.add(handler);
-        return this;
-    }
-
-    public TCPFactory<B> onDisconnect(Callback<Void, Void> handler) {
-        this.exitHandlers.add(handler);
-        return this;
-    }
-
-    public TCPFactory<B> onConnect(Callback<Void, Void> handler) {
-        this.connectHandlers.add(handler);
-        return this;
-    }
-
-    public void exit() {
-        this.exitHandlers.forEach((eh) -> eh.handle(null));
-    }
-
-    public void connect() {
-        this.connectHandlers.forEach((th) -> th.handle(null));
-    }
-
-    public void error(Exception e) {
-        this.errorHandlers.forEach((eh) -> eh.handle(e));
-    }
-
-    public void data(B t) {
-        this.handlers.forEach((th) -> th.handle(t));
     }
 }
